itest for marketsv1 with 8MiB deal
